Add optional formatted print of processed options
Triggers only when -d is chosen.  Also started debug print of option
processing, if the -d flag is passed.

Also, fixed that -r was marked as taking options, when it really
shouldn't.  This was causing it to gobble the next options.

Also, we weren't even reading in the number-of-teeth values somehow.  That
is also fixed.