Fix math on continuity flip in genPointList()
Previously we were multiplying an unsigned char by -1.0, casting back to
an unsigned char, and magically getting output that was what we expected.
While this works on the compiler we're using, it was unlikely to be
portable, and is just confusing to read.

Switched to the sane method for inverting an unsigned char that has a
signal centered around some value:
-(x - offset) + offset = -x + 2*offset

Also added comments to make it clearer what that function is doing.

Some stale, commented out code was also removed from the file.